cannon: Move vmtype to load_elf since it is no longer shared.
